+++ packages/atomic/src/components/commerce/atomic-commerce-interface/atomic-commerce-interface.tsx
@@ -12,10 +12,9 @@ import {
   ProductListing,
   Context,
   buildContext,
-  buildSearchSummary,
-  buildListingSummary,
-  SearchSummary,
-  ListingSummary,
+  Summary,
+  SearchSummaryState,
+  ProductListingSummaryState,
 } from '@coveo/headless/commerce';
 import {
   Component,
@@ -48,7 +47,7 @@ import {
 import {getAnalyticsConfig} from './analytics-config';
 import {AtomicCommerceStore, createAtomicCommerceStore} from './store';
 
-const FirstSearchExecutedFlag = 'firstSearchExecuted';
+const FirstRequestExecutedFlag = 'firstRequestExecuted';
 
 export type CommerceInitializationOptions = CommerceEngineConfiguration;
 export type CommerceBindings = CommonBindings<
@@ -76,10 +75,9 @@ export class AtomicCommerceInterface
 {
   private urlManager!: UrlManager;
   private searchOrListing!: Search | ProductListing;
-  private summary!: SearchSummary | ListingSummary;
+  private summary!: Summary<SearchSummaryState | ProductListingSummaryState>;
   private context!: Context;
   private unsubscribeUrlManager: Unsubscribe = () => {};
-  private unsubscribeSearchStatus: Unsubscribe = () => {};
   private unsubscribeSummary: Unsubscribe = () => {};
   private initialized = false;
   private store: AtomicCommerceStore;
@@ -115,8 +113,10 @@ export class AtomicCommerceInterface
 
   /**
    * the commerce interface language.
+   *
+   * Will default to the value set in the Headless engine context if not provided.
    */
-  @Prop({reflect: true}) public language = 'en';
+  @Prop({reflect: true, mutable: true}) public language?: string;
 
   /**
    * The commerce interface headless engine.
@@ -181,7 +181,7 @@ export class AtomicCommerceInterface
   }
 
   public connectedCallback() {
-    this.store.setLoadingFlag(FirstSearchExecutedFlag);
+    this.store.setLoadingFlag(FirstRequestExecutedFlag);
     this.i18nClone = this.i18n.cloneInstance();
     this.i18n.addResourceBundle = (
       lng: string,
@@ -209,6 +209,9 @@ export class AtomicCommerceInterface
     if (!this.commonInterfaceHelper.engineIsCreated(this.engine)) {
       return;
     }
+    if (!this.language) {
+      return;
+    }
 
     this.context.setLanguage(this.language);
     this.commonInterfaceHelper.onLanguageChange();
@@ -221,7 +224,6 @@ export class AtomicCommerceInterface
 
   public disconnectedCallback() {
     this.unsubscribeUrlManager();
-    this.unsubscribeSearchStatus();
     this.unsubscribeSummary();
     window.removeEventListener('hashchange', this.onHashChange);
   }
@@ -368,44 +370,42 @@ export class AtomicCommerceInterface
       this.type === 'product-listing'
         ? buildProductListing(this.engine!)
         : buildSearch(this.engine!);
-
-    this.unsubscribeSearchStatus = this.searchOrListing.subscribe(() => {
-      if (
-        !this.searchOrListing.state.isLoading &&
-        this.store.hasLoadingFlag(FirstSearchExecutedFlag)
-      ) {
-        this.store.unsetLoadingFlag(FirstSearchExecutedFlag);
-      }
-    });
   }
 
   private initSummary() {
-    this.summary =
-      this.type === 'product-listing'
-        ? buildListingSummary(this.engine!)
-        : buildSearchSummary(this.engine!);
+    this.summary = this.searchOrListing.summary();
 
     this.unsubscribeSummary = this.summary.subscribe(() => {
-      const {firstSearchExecuted, hasProducts, hasError} = this.summary.state;
-      const hasNoProductsAfterInitialSearch =
-        firstSearchExecuted && !hasError && !hasProducts;
+      const {firstRequestExecuted, hasProducts, hasError} = this.summary.state;
+      const hasNoProductsAfterInitialQuery =
+        firstRequestExecuted && !hasError && !hasProducts;
 
       this.host.classList.toggle(
         noProductsSelector,
-        hasNoProductsAfterInitialSearch
+        hasNoProductsAfterInitialQuery
       );
 
       this.host.classList.toggle(errorSelector, hasError);
 
       this.host.classList.toggle(
         firstSearchExecutedSelector,
-        firstSearchExecuted
+        firstRequestExecuted
       );
+
+      if (firstRequestExecuted) {
+        this.store.unsetLoadingFlag(FirstRequestExecutedFlag);
+      }
     });
   }
 
   private initContext() {
     this.context = buildContext(this.engine!);
+  }
+
+  private initLanguage() {
+    if (!this.language) {
+      this.language = this.context.state.language;
+    }
   }
 
   private updateHash() {
@@ -433,6 +433,7 @@ export class AtomicCommerceInterface
     this.initSummary();
     this.initUrlManager();
     this.initContext();
+    this.initLanguage();
     this.initialized = true;
   }