+++ packages/atomic/src/components/commerce/facets/atomic-commerce-facets/e2e/fixture.ts
@@ -1,0 +1,18 @@
+import {test as base} from '@playwright/test';
+import {
+  AxeFixture,
+  makeAxeBuilder,
+} from '../../../../../../playwright-utils/base-fixture';
+import {FacetsPageObject} from './page-object';
+
+interface TestFixture {
+  facets: FacetsPageObject;
+}
+
+export const test = base.extend<TestFixture & AxeFixture>({
+  makeAxeBuilder,
+  facets: async ({page}, use) => {
+    await use(new FacetsPageObject(page));
+  },
+});
+export {expect} from '@playwright/test';