+++ packages/headless/src/controllers/commerce/search/headless-search.test.ts
@@ -1,5 +1,10 @@
 import {configuration} from '../../../app/common-reducers';
 import {contextReducer as commerceContext} from '../../../features/commerce/context/context-slice';
+import {
+  pagePrincipalSelector,
+  perPagePrincipalSelector,
+  totalEntriesPrincipalSelector,
+} from '../../../features/commerce/pagination/pagination-selectors';
 import {searchSerializer} from '../../../features/commerce/parameters/parameters-serializer';
 import {queryReducer as commerceQuery} from '../../../features/commerce/query/query-slice';
 import {restoreSearchParameters} from '../../../features/commerce/search-parameters/search-parameters-actions';
@@ -8,6 +13,10 @@ import * as SearchActions from '../../../features/commerce/search/search-actions
 import {
   activeParametersSelector,
   enrichedParametersSelector,
+  enrichedSummarySelector,
+  errorSelector,
+  isLoadingSelector,
+  numberOfProductsSelector,
   requestIdSelector,
   responseIdSelector,
 } from '../../../features/commerce/search/search-selectors';
@@ -66,6 +75,13 @@ describe('headless search', () => {
       restoreActionCreator: restoreSearchParameters,
       activeParametersSelector,
       enrichParameters: enrichedParametersSelector,
+      isLoadingSelector,
+      errorSelector,
+      pageSelector: pagePrincipalSelector,
+      perPageSelector: perPagePrincipalSelector,
+      totalEntriesSelector: totalEntriesPrincipalSelector,
+      numberOfProductsSelector,
+      enrichSummary: enrichedSummarySelector,
     });
   });